requirements.txt
requests 
pillow
pytesseract
beautifulsoup4